feat : VO 메서드 생성 및 코드 포맷팅 작업.
