--- src/main/java/com/next/step/step3/domain/CarName.java
@@ -1,14 +1,0 @@
-package com.next.step.step3.domain;
-
-public class CarName {
-
-    private String name;
-
-    public CarName(String name) {
-        this.name = name;
-    }
-
-    public String carName() {
-        return name;
-    }
-}
